+++ jans-core/service/src/main/java/io/jans/config/link/Configuration.java
@@ -1,0 +1,16 @@
+/*
+ * Janssen Project software is available under the Apache License (2004). See http://www.apache.org/licenses/ for full text.
+ *
+ * Copyright (c) 2020, Janssen Project
+ */
+
+package io.jans.config.link;
+
+/**
+ * base interface for all oxAuth configurations
+ *
+ * @author Yuriy Movchan
+ * @version 04/12/2017
+ */
+public interface Configuration {
+}