+++ jans-link/model/src/main/java/io/jans/link/event/CacheRefreshEvent.java
@@ -1,4 +1,0 @@
-package io.jans.link.event;
-
-public class CacheRefreshEvent {
-}